Various alerting rule cleanups/fixes etc.
- Changed all alert names to CamelCase (@bjk-gitlab asked for this)
- Improved/fixed some alert conditions and messages
- Try to preserve a tad more labels in expressions (when possible)

Unfortunately the Slack notification templates in Alertmanager rely on
there being only one "title" and "description" value, because they use
the title/description from CommonAnnotations (so those are only present
if they are the same between all alerts resulting from one alerting
rule). So I kept appropriate aggregations and topk-ing in some of the
alerts to not break that assumption.