Add safeguards to prevent setting an invalid environment
If the required environment variables are not set the functions script
will fail. In the same spirit, the default values will not be set if
the folder or files do not exists.

This is so to prevent having a partially set environment.

Additionally, all the functions that fail will return 1 to set the right
return code.