Increase threshold for connection amount alerts
Pgbouncer is set to allow up to 250 connections, which PostgreSQL's
"max_connections" being set to 300. Using a threshold of 0.75 we would
fire alerts if more than 225 connections are used. With 25 connections
to spare (and pgbouncer being able to queue connections) this is a bit
of an overkill.

To reduce the noise I'm increasing the threshold to 0.85 meaning we
won't alert unless 255 or more connections are being used (this includes
the 5 reserve connections we have).

In other words, up until 250 connections we are perfectly fine and so
there's no need to alert until we hit that threshold.