Moving towards simplifying the alerts.
In future we want to pagerduty on critical alerts, not on alerts that
are set to `pager=pagerduty`. This is a first step in that direction.
Pagerduty alerts are critical alerts. Critical alerts are pagerduty
alerts.

As a next step, we can simply start routing alerts to pagerduty when
severity is critical and drop the pager attribute altogether.