Alert on low ZFS space
Add zpool devices to the recording rule regex that is used in filesystem
space alerting rules. This couples the rule to the naming scheme used in
the gitlab_disks cookbook, which is not ideal but it has to be coupled
to something.

"tank/dataset" is exclusive of the ZFS reservation, which is 10% of the
available disk space. This will cause the alerts to fire at lower
utilisations for ZFS than other filesystems, and this is desirable. ZFS
has performance issues near full capacity, hence the reservation.

Did not add ZFS to any inode alerts, as ZFS creates metadata dynamically
and does not have a pre-provisioned inode space like ext4.