Adjust SLO expectations for git
Recently we changed the way we performed health checks, with the effect
that they're no longer routed through workhorse (see
https://gitlab.com/gitlab-com/gl-infra/production/issues/1268 for more
details). The fact that these lightweight requests no longer contribute
to our latency averages means it has gone up. Currently we're receiving
intermittent alerts for git despite the service being healthy, so we're
adjusting its SLO in accordance to what we're observing. It's possible
we might need to adjust other SLOs in the near future