Add sidekiq worker saturation monitoring and visualization
This change adds a resource saturation metric for sidekiq workers.

When sidekiq workers become saturated, an new jobs that are added will
be queued. This is acceptable over the short term, but long term should
be avoided.

By adding a resource saturation metric, we also get long term capacity
planning on the Platform Triage dashboard.

This change also includes a visualization for the metric, on the
sidekiq main dashboard, as well as the priority detail dashboard.