This new "How To" doc is a consolidation of research and incident analyses involving Patroni failovers, Consul/Patroni interactions, and PgBouncer CPU saturation and workload apportioning.
The "Quick orientation" section gives an overview, aiming to be helpful to an on-call engineer (EOC) who may not be familiar with all the components of this subsystem. Similarly the "Where do the components run?" section aims mainly to help the EOC, but can also be helpful to anyone wanting to learn more by observing a live system or by reviewing the relevant sections of Chef code.

The "Links to external docs" section covers a short list of the most relevant external docs. Throughout the rest of the doc, more specific contextual links are included for external documentation, relevant GitLab.com Issues, and source code.

The "Quick reference commands" section is a collection of one-liners, mostly used for exploring/inspecting state, not changing state, so they are safe to use on healthy environments too unless otherwise indicated.

The remainder of this doc is organized as a question/answer list to break up the material into more easily digestible chunks and allow easier navigation. The material flows linearly building on previous topics, but each section aims to give enough context to make sense by itself.