Increase alert breach time for gitaly latency
We currently have many flappy Gitaly latency alerts, and it's normal for
the on-call to find that latency has already returned to normal before
the alert fired, and also that the repositories responsible are rarely
abusive.

In the long run, we want to improve Gitaly latency through caching (even
if this isn't performed at Gitaly itself, fewer requests will hit Gitaly).