Remove the transaction around the project import, setting the idle timeout
It is problematic, as it forces the entire import into a single transaction where it would not natively be one.  For large imports, this has severe and increasing-over-time impact on the primary database as it keeps track of all the transaction data/history.

Work done in the rapid action response to a customer import (https://docs.google.com/document/d/1EVBZHRWjd-sPy4Y410ZuCoobqELJRd1wbej5E2SR-SY) should solve this properly and permanently