Alert on Cloudflare exporter scrape errors
Due to the way Cloudflare expose their analytics metrics, this exporter
doesn't gather them at scrape time, but in a background loop. As such it
does not report scrape errors to prometheus when these background
scrapes are failing. We must monitor for these background scrape errors
in order to know when our Cloudflare metrics are not being updated.