Allow the saturation alert dashboard to have a default 'type' other than web
Some of the saturation alert dashboards do not make sense with a default type of 'web', and it can be confusing when finding these via
means other than an alert link, when they show no data by default.  And sometimes finding the correct 'type' to choose is challenging.

This commit adds the capability for a saturation dashboard to supply a default 'type' value, and uses that on a number of dashboards that need it (there may be more, these are the ones I could easily identify)