Update pgbouncer documentation
Mainly, we're not using omnibus on our patroni or pgbouncer nodes
anymore. This means we can't rely on gitlab-ctl, and some paths change.
These changes lay the ground for further updates we're planning on
https://gitlab.com/gitlab-com/gl-infra/infrastructure/-/issues/10379 and
https://gitlab.com/gitlab-com/gl-infra/infrastructure/-/issues/10425. We
should keep an eye on returning to omnibus-gitlab land once patroni is
integrated with it (see https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/5115)

Also, we rely more solidly on consul now, so reflect that.