Limit SLO lines on redis memory saturation to just one
Without the environment, we get metrics from every environment (repeated, same value).  The max() whittles the per-environment values down to one; without that we get one from each prometheus shard (main, app, db, and GKE); the values "should" be the same in all monitoring shards in an environment, so max() is an arbitrary choice.