Respect the stage param for saturation panels
The service_dashboard template can be passed a `stage` parameter. By
default, it's set to `$stage` so that it can be switched between `main`
and `cny`. For pgbouncer and patroni there's only the `main` stage, so
their dashboards specify this parameter to be hardcoded to `main`.
Because the saturation panels weren't respecting this parameter, they
were still being rendered with `$main`, which make them not function for
those two services.