Add alert CICDJobsStuckInDockerPull
This is for https://gitlab.com/gitlab-com/gl-infra/infrastructure/-/issues/10959

This is a fairly rudimentary alert that will fire when the number of CI jobs
in state `docker_pulling_image` exceeds the number of CI jobs in state `docker_run`
for more than 5 minutes. This is a symptom that there is connectivity issues with
docker hub.