Ignore thanos labels in recording rule resolution
When evaluating whether to use a recording rule instead of a metric,
the metrics catalog may encounter labels that were applied after the recording
rule was evaluated.

Before this change, the logic would dictate the the recording rule could not be
used since the label had not been applied to the expression.

This change allows the evaluation to ignore thanos level rules, allowing
the recording rule to be applied as expected, while safely allowing thanos
routing labels to be applied.