Corrects the `for` on KubeReplicasSetPodMismatch to match the text
Empirically 5 minutes is quite short (it can take several minutes for large numbers of pods to spin up when required, e.g. for sidekiq) and it's not hard for is to be one or two shy of the required number without there being a massive problem.  I'm inclined to believe the text was the intended time frame, both because of the effort required to put that in, and because it will make the alert much less twitchy.