Add recording rules for http_requests_total to other services
Because we do the metric replacement (swapping the 'raw' metric with the
recording rule) based on the metric name and labels, when we add
a recording rule for `http_requests_total` to one service, we implicitly
add it to all services using that rule.

This adds it explicitly to all services using that metric to make that a
little less confusing. It will result in increased time evaluating the
rules, because it will evaluate each rule four times - once for each
service.