Make pagerduty alerts repeat every 2h, not every 8h
By default the repeat_interval is 8hr.  However this is incompatible with the default 6hr expiry/auto-resolution in PagerDuty, leading to confusion on the part of the on-calls as alerts 'Resolve' then fire again 2 hours later when AlertManager sends the next notification to PD.

It needs to be stricly <6h , but we should consider that very occasionally there are failures sending from AM to PD.  And if it's 3h, then if the mid-event notification is lost, and their is a *slight* delay in sending the second, we could get a resolve + re-fire, again being confusing.  2h keeps the rate of notifications low, but gives a good chance of at least 1 making it to PD.